Refactored Address and Name into Attribute package
