+++ src/test/java/seedu/loyaltylift/model/customer/AddressTest.java
@@ -6,6 +6,8 @@ import static seedu.loyaltylift.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.loyaltylift.model.attribute.Address;
+
 public class AddressTest {
 
     @Test
